% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fn.R
\name{analysis}
\alias{analysis}
\title{code to analyze single dataset of n observations using our list of
estimators.}
\usage{
analysis(
  dat,
  radius = 10,
  min_sample = 8,
  n_sentinel = 20,
  startnum = 50,
  endnum = 100,
  include_OLS = TRUE,
  include_BINDING = FALSE,
  include_FRONTIER = FALSE,
  include_GP = FALSE,
  include_GP_RES = FALSE,
  include_loess = FALSE
)
}
\arguments{
\item{dat}{data to be analyzed}

\item{radius}{Parameter to loess estimator.}

\item{min_sample}{Parameter to loess estimator.}

\item{n_sentinel}{number of sentinels per side}

\item{startnum}{Parameter to gaussian process estimator}

\item{endnum}{Parameter to gaussian process estimator.}

\item{include_OLS}{for OLS regression}

\item{include_BINDING}{for binding score}

\item{include_FRONTIER}{for pooled frontier}

\item{include_GP}{for GPR}

\item{include_GP_RES}{for residualized GPR}

\item{include_loess}{for loess}
}
\description{
code to analyze single dataset of n observations using our list of
estimators.
}
