% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_AFE_SE}
\alias{calc_AFE_SE}
\title{Calculate the BATE SE}
\usage{
calc_AFE_SE(GP_res, weight)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in.}

\item{weight}{Vector of weights.}
}
\description{
Use calc_Ybar_Var function on both the treated and control sides.
Weight the sentinels on the boundary as given by 'weight'.
}
