% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_code.R
\name{calc_edge_weight}
\alias{calc_edge_weight}
\title{Calculate edge weights.}
\usage{
calc_edge_weight(sents)
}
\arguments{
\item{sents}{The sentinels to be analyzed.}
}
\value{
A tibble with the total weight, and the weight on each edge
}
\description{
Given a list of sentinels, calculate the total weight on each edge
and the total weight overall.
}
