% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{calc_true_effects}
\alias{calc_true_effects}
\title{Calculate true effects}
\usage{
calc_true_effects(rating1, rating2, E.true.T, rMean, rSigma, tx.func)
}
\arguments{
\item{rating1}{Running variable 1.}

\item{rating2}{Running variable 2.}

\item{E.true.T}{Defined as tx.func(r1, r2) + eff.err.}

\item{rMean}{Defined as c( mu1 - cut1.old, mu2 - cut2.old), where the old cuts
are the original running variable cutoffs.}

\item{rSigma}{Defined as matrix(c(sd(truerating1) ^ 2, cov.r1r2, cov.r1r2, sd(truerating2) ^ 2), c(2, 2)).}

\item{tx.func}{The treatment function.}
}
\description{
Calculate true effects assuming the running variable distribution
is bivariate normal with passed mean and covariance.
I.e., (r1, r2) ~ MVN( rMean, rSigma )
}
