% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{calc_true_sentinel_weights}
\alias{calc_true_sentinel_weights}
\title{Calculate weights at sentinels}
\usage{
calc_true_sentinel_weights(rating1, rating2, rMean, rSigma)
}
\arguments{
\item{rating1}{Running variable 1.}

\item{rating2}{Running variable 2.}

\item{rMean}{Defined as c( mu1 - cut1.old, mu2 - cut2.old), where the old cuts
are the original running variable cutoffs.}

\item{rSigma}{Defined as matrix(c(sd(truerating1) ^ 2, cov.r1r2, cov.r1r2, sd(truerating2) ^ 2), c(2, 2)).}
}
\value{
List of weights, one for each sentinel.  Sums to 1.
}
\description{
Given the true mean and covariance, calculate weights of a passed
sequence of (r1,r2) pairs.
}
