% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_tx_curve}
\alias{calc_tx_curve}
\title{Calculate the treatment curve along the given sentinels.}
\usage{
calc_tx_curve(sentinels, data, method, startnum, endnum)
}
\arguments{
\item{sentinels}{df of sentinels}

\item{data}{data to be analyzed, passed}

\item{method}{What type of GP to call, passed}

\item{startnum}{Parameter to gaussian process estimator}

\item{endnum}{Parameter to gaussian process estimator}
}
\description{
This function fits the GP twice, once for each dataset.
}
\details{
Middle function - divides data into Treated & Control before laGP prediction
}
