% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate weights of the sentinels.}
\usage{
calc_weights(x, y, data)
}
\arguments{
\item{x}{location of sentinels, sentinels$rating1}

\item{y}{location of sentinels, sentinels$rating2}

\item{data}{data to be analyzed}
}
\description{
Here we fit a multivariate normal and then do weights as density
given the assumed multivariate normal.
}
