% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calculate_precision_BATE}
\alias{calculate_precision_BATE}
\title{This implements equation 13 from Maxime paper, and gives the
inverse variance weighted BATE with associated estimated standard
error.}
\usage{
calculate_precision_BATE(GP_res)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in}
}
\description{
NOTE: The code below does not use this function since it calculates
weights via calculate_precision_weights() and then just uses those
weights as any other weights for the following calculations.
}
