% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calculate_precision_weights}
\alias{calculate_precision_weights}
\title{This implements equation 13 from Maxime paper to get sentinal
weights that correspond to the inverse variance weighting.}
\usage{
calculate_precision_weights(GP_res)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in}
}
\description{
This implements equation 13 from Maxime paper to get sentinal
weights that correspond to the inverse variance weighting.
}
