% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{drop_low_weight_sentinels}
\alias{drop_low_weight_sentinels}
\title{Given a list of sentinels, drop those with very low weight (using
the weight column).  Will drop smallest sentinels up to the drop
proportion, but never exceed that threshold.}
\usage{
drop_low_weight_sentinels(GP_res, drop_prop = 0.01)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in}

\item{drop_prop}{proportion of weight that}
}
\description{
Given a list of sentinels, drop those with very low weight (using
the weight column).  Will drop smallest sentinels up to the drop
proportion, but never exceed that threshold.
}
