% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{gaussianp}
\alias{gaussianp}
\title{Use Gaussian process regression}
\usage{
gaussianp(
  sampdat,
  n_sentinel = 20,
  method = c("new", "aGP", "laGP"),
  startnum,
  endnum,
  fixed_sent = FALSE
)
}
\arguments{
\item{sampdat}{Data to analyze.}

\item{n_sentinel}{The number of sentinels per side.}

\item{method}{What type of GP to call - newGP, aGP, or laGP (Use
'new' generally, although it is more time intensive, others are
approximations).}

\item{startnum}{Parameter to Gaussian process estimator.}

\item{endnum}{Parameter to Gaussian process estimator.}

\item{fixed_sent}{TRUE if user wants to use fixed sentinel locations between 0-4 on each running variable.}
}
\description{
Given a two-dimensional dataset with two cutpoints, estimate treatment effects
along the boundary.
This method uses a Gaussian process regression on the treated and
control units to estimate a treatment impact along the boundary
defined by having a 0 score in at least one of the two ratings.
This function will label units as treated or not, based on the
(rating1, rating2) values.
}
