% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{gen.2RRDD}
\alias{gen.2RRDD}
\title{function gen.2RRDD}
\usage{
gen.2RRDD(
  n,
  s = 1,
  mu1 = 0,
  mu2 = 0,
  sigma.r1 = 1,
  sigma.r2 = sigma.r1,
  rho.r1r2 = 0,
  cut1.quantile = NULL,
  cut2.quantile = NULL,
  cut1.value = NULL,
  cut2.value = NULL,
  param.Y0 = c(0, 0, 0, 0, 0, 0),
  tx.func = function(r1, r2) {
     0.4
 },
  sigma.E = 0,
  sigma.noise = 1,
  data.type = "observed",
  seed = NULL
)
}
\arguments{
\item{n}{= sample size}

\item{s}{= number of samples}

\item{mu1}{= mean of rating 1}

\item{mu2}{= mean of rating 2}

\item{sigma.r1}{= sd of true rating 1}

\item{sigma.r2}{= sd of true rating 2}

\item{rho.r1r2}{= correlation between true ratings}

\item{cut1.quantile}{= quantile for cutting rating 1}

\item{cut2.quantile}{= quantile for cutting rating 2}

\item{cut1.value}{= value for cutting rating 1 (if no quantile
passed)}

\item{cut2.value}{= value for cutting rating 2 (if no quantile
passed).}

\item{param.Y0}{= parameters for fn relating ratings and Y0}

\item{sigma.E}{The sd of random error in effects (assumed same for
all params)}

\item{sigma.noise}{The sd of error term in surface response model}

\item{data.type}{= "full","observed",or "none" (no data returned -
just parameter values)}

\item{param.eff.T}{= parameters for treatment impact as function of
ratings for full model with terms r1,r2,r1*r2,r1^2,r2^2}
}
\value{
: simdat = either full or observed dataset and some
descriptive info such as parameter values for observed data.
}
\description{
Generate fake data following general structure of Porter et al, as
adjusted by An at al (2024)
}
