% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{gen_dat_sim}
\alias{gen_dat_sim}
\title{This function generates multiple datasets at once, given the
simulation type, cut scores, sample size.}
\usage{
gen_dat_sim(
  sim = 1,
  rho = 0,
  cut1 = 0.5,
  cut2 = 0.5,
  seed = NULL,
  n = 5000,
  s = 500
)
}
\arguments{
\item{sim}{ID for the simulation type, following Porter et al.}

\item{n}{Number of samples per set}

\item{s}{Number of datasets}
}
\description{
Passed parameters allow for changing the data generating process.
}
