% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussiantwodRDD.R
\name{get_cov_matrix_side}
\alias{get_cov_matrix_side}
\title{Retrieve sentinels' covariance matrix}
\usage{
get_cov_matrix_side(GP_res, tx = NULL)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in.}

\item{tx}{Needed to identify Sigmas: treated = 1, control = 0.}
}
\description{
Pull the covariance matrix of the sentinels out of the GP result
object (treatment or control side)
}
