% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loesstwodRDD.R
\name{loess2DRDD}
\alias{loess2DRDD}
\title{Generate loess estimates}
\usage{
loess2DRDD(sampdat, radius, min_sample, n_sentinel = 20, sentinels)
}
\arguments{
\item{sampdat}{Data to be analyzed.}

\item{radius}{Radius size around each point (we use half a SD).}

\item{min_sample}{Minimum number of data points within each radius.}

\item{n_sentinel}{Number of sentinels per side of the boundary.}

\item{sentinels}{The sentinels used for prediction, with low-weight sentinels already dropped.}
}
\description{
Given data and a list of sentinels, calculate predicted Y0 and Y1, and then
take the difference as the estimated treatment effect at each sentinel
}
