% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{make_sentinels}
\alias{make_sentinels}
\title{Generate a grid of sentinels along the treatment boundary (assumed
to be where one of ratings is 0)}
\usage{
make_sentinels(dat, n_sentinel, fixed_sent = FALSE)
}
\arguments{
\item{dat}{data to be analyzed}

\item{n_sentinel}{number of sentinels per side}

\item{fixed_sent}{fixed sentinel locations between 0-4 on each RV}
}
\description{
Generate a grid of sentinels along the treatment boundary (assumed
to be where one of ratings is 0)
}
