% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussiantwodRDD.R
\name{params}
\alias{params}
\title{Return ML parameters of the GP model}
\usage{
params(gp_res)
}
\arguments{
\item{gp_res}{Result from gaussianp.}
}
\description{
This function extracts the parameters of the Gaussian process,
given the result from gaussianp
}
