% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_items.R
\docType{package}
\name{twodrdd}
\alias{twodrdd-package}
\alias{twodrdd}
\title{\code{twodrdd} package}
\description{
This package allows the user to use nonparametric response surface methods as well as methods using local linear regression to analyze two- dimensional regression discontinuity designs. These methods use the rddapp and laGP packages extensively; code is also provided to generate data sets that was originally shared by Kristin Porter.
}
\references{
An et. al (2024) Our paper on using Gaussian process regression as a nonparametric surface response method for two-dimensional RDDs in education.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lily-an1/twodrdd}
  \item Report bugs at \url{https://github.com/lily-an1/twodrdd/issues}
}

}
\author{
\strong{Maintainer}: Lily An \email{lily_an@g.harvard.edu} (\href{https://orcid.org/0000-0002-1931-1533}{ORCID})

Authors:
\itemize{
  \item Luke Miratrix (\href{https://orcid.org/0000-0002-0078-1906}{ORCID})
  \item Zach Branson (\href{https://orcid.org/0000-0001-9496-9979}{ORCID})
}

}
