% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fn.R
\name{analysis}
\alias{analysis}
\title{Two-dimensional RDD methods}
\usage{
analysis(
  dat,
  cut1 = 0,
  cut2 = 0,
  min_sample = 8,
  n_sentinel = 20,
  startnum = 50,
  endnum = 100,
  include_OLS = TRUE,
  include_BINDING = FALSE,
  include_FRONTIER = FALSE,
  include_GP = FALSE,
  include_GP_RES = FALSE,
  include_loess = FALSE
)
}
\arguments{
\item{dat}{Data to be analyzed.}

\item{cut1}{Cutoff value for running variable 1. Centered at 0 in the simulation's data generating processes.}

\item{cut2}{Cutoff value for running variable 2. centered at 0 in the simulation's data generating processes.}

\item{min_sample}{Parameter to loess estimator.}

\item{n_sentinel}{The number of sentinels per side.}

\item{startnum}{Parameter to the gaussian process estimator.}

\item{endnum}{Parameter to the gaussian process estimator.}

\item{include_OLS}{TRUE to include OLS regression.}

\item{include_BINDING}{TRUE to include binding score.}

\item{include_FRONTIER}{TRUE to include pooled frontier.
The results parameter is labeled the AFE. AFE1 and AFE2 are the effects specific to
each running variable boundary.}

\item{include_GP}{TRUE to include Gaussian process regression.}

\item{include_GP_RES}{TRUE to include residualized Gaussian process regression.}

\item{include_loess}{TRUE to include loess.}
}
\description{
Code to analyze single dataset of n observations using our list of
estimators. The treated area is the top right quadrant.
}
