% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_AFE_SE}
\alias{calc_AFE_SE}
\title{Calculate the Standard Error for the Average Frontier Effect}
\usage{
calc_AFE_SE(GP_res, weight)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in}

\item{weight}{vector of weights}
}
\description{
Weight the sentinels on the frontier as given by 'weight'.
}
