% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_one_point_GP}
\alias{calc_one_point_GP}
\title{Fit gaussian process and use it to predict for each (x,y) pair
passed.}
\usage{
calc_one_point_GP(
  sentinels,
  data,
  method = c("new", "aGP", "laGP"),
  startnum = NULL,
  endnum = NULL,
  prefix = "var."
)
}
\arguments{
\item{sentinels}{Dataframe holding rating1,rating2, which are the}

\item{data}{data to be analyzed, passed}

\item{method}{What type of GP to call, passed}

\item{startnum}{Parameter to gaussian process estimator}

\item{endnum}{Parameter to gaussian process estimator}

\item{prefix}{used in Sigma column names
x,y locations of the sentinels.}
}
\description{
Fit gaussian process and use it to predict for each (x,y) pair
passed.
}
