% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{calc_true_effects}
\alias{calc_true_effects}
\title{Calculate true effects assuming the running variable distribution
is bivariate normal with passed mean and covariance.
I.e., (r1, r2) ~ MVN( rMean, rSigma )}
\usage{
calc_true_effects(rating1, rating2, E.true.T, rMean, rSigma, tx.func)
}
\description{
Calculate true effects assuming the running variable distribution
is bivariate normal with passed mean and covariance.
I.e., (r1, r2) ~ MVN( rMean, rSigma )
}
