% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{calc_true_sentinel_weights}
\alias{calc_true_sentinel_weights}
\title{Given the true mean and covariance, calculate weights of a passed
sequence of (r1,r2) pairs.}
\usage{
calc_true_sentinel_weights(rating1, rating2, rMean, rSigma)
}
\value{
List of weights, one for each sentinel.  Sums to 1.
}
\description{
Given the true mean and covariance, calculate weights of a passed
sequence of (r1,r2) pairs.
}
