% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_tx_curve}
\alias{calc_tx_curve}
\title{Calculate the treatment curve along the given sentinels.}
\usage{
calc_tx_curve(sentinels, data, method, startnum, endnum)
}
\arguments{
\item{sentinels}{Data frame of the sentinels.}

\item{data}{Data to be analyzed, passed.}

\item{method}{What type of GP to call, passed.}

\item{startnum}{Parameter to Gaussian process estimator.}

\item{endnum}{Parameter to Gaussian process estimator.}
}
\description{
This function fits the GP twice, once for each dataset.
Middle function - divides data into Treated & Control before Gaussian process prediction.
}
