% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{calc_weights}
\alias{calc_weights}
\title{Sentinel weights}
\usage{
calc_weights(x, y, data)
}
\arguments{
\item{x}{The location of sentinels, sentinels$rating1.}

\item{y}{The location of sentinels, sentinels$rating2.}

\item{data}{The data to be analyzed.}
}
\description{
Calculate weights of the sentinels.
Here we fit a multivariate normal and then do weights as density
given the assumed multivariate normal.
}
