% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussiantwodRDD.R
\name{calculate_precision_weights}
\alias{calculate_precision_weights}
\title{Precision weights}
\usage{
calculate_precision_weights(GP_res)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in.}
}
\description{
This implements equation 13 from Rischard et al. (2021) to get sentinel
weights that correspond to the inverse variance weighting.
}
