% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{gen.2RRDD}
\alias{gen.2RRDD}
\title{Generate fake data}
\usage{
gen.2RRDD(
  n,
  s = 1,
  mu1 = 0,
  mu2 = 0,
  sigma.r1 = 1,
  sigma.r2 = sigma.r1,
  rho.r1r2 = 0,
  cut1.quantile = NULL,
  cut2.quantile = NULL,
  cut1.value = NULL,
  cut2.value = NULL,
  param.Y0 = c(0, 0, 0, 0, 0, 0),
  tx.func = function(r1, r2) {
     0.4
 },
  sigma.E = 0,
  sigma.noise = 1,
  data.type = "observed",
  seed = NULL
)
}
\arguments{
\item{n}{= Sample size.}

\item{s}{= Number of samples.}

\item{mu1}{= Mean of running variable 1.}

\item{mu2}{= Mean of running variable 2.}

\item{sigma.r1}{= Standard deviation of true running variable 1.}

\item{sigma.r2}{= Standard deviation of true running variable 2.}

\item{rho.r1r2}{= Correlation between true running variables.}

\item{cut1.quantile}{= Quantile for cutting running variable 1.}

\item{cut2.quantile}{= Quantile for cutting running variable 2.}

\item{cut1.value}{= Value for cutting running variable 1 (if no quantile
passed).}

\item{cut2.value}{= Value for cutting running variable 2 (if no quantile
passed).}

\item{param.Y0}{= Parameters for the function relating running variables and Y0.}

\item{tx.func}{The treatment function for true impacts,}

\item{sigma.E}{The standard deviation of random error in effects (assumed same for
all params).}

\item{sigma.noise}{The standard deviation of the error term in surface response model.}

\item{data.type}{= Either "full" or "observed". Observed contains ID,
running variable1, running variable2, T, Y).}

\item{seed}{Set seed parameter.}
}
\value{
: simdat = Either full or observed dataset and some
descriptive info such as parameter values for observed data.
}
\description{
Generate fake data following general structure of Porter et al. (2017), as
adjusted by An et al. (2024)
}
