% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{gen_dat_sim}
\alias{gen_dat_sim}
\title{Create simulation data}
\usage{
gen_dat_sim(
  sim = 1,
  rho = 0,
  cut1 = 0.5,
  cut2 = 0.5,
  seed = NULL,
  n = 5000,
  s = 500
)
}
\arguments{
\item{sim}{ID for the simulation type, following Porter et al. (2017).}

\item{rho}{Correlation between running variables.}

\item{cut1}{Percentile at which running variable 1's cutoff is.}

\item{cut2}{Percentile at which running variable 2's cutoff is.}

\item{seed}{Set seed parameter.}

\item{n}{Number of samples per set.}

\item{s}{Number of datasets.}
}
\description{
This function generates multiple datasets at once, given the
simulation type, cut scores, sample size.
Passed parameters allow for changing the data generating process.
Modified from Kristin Porter's code shared Feb 2022 for Porter et al. (2017).
}
