% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianGeoRDD.R
\name{get_cov_matrix_tx}
\alias{get_cov_matrix_tx}
\title{Add covariance matrices}
\usage{
get_cov_matrix_tx(GP_res)
}
\arguments{
\item{GP_res}{Gaussian process regression result to be passed in.}
}
\description{
This is Sigma_b|Y (Equation 6 in Rischard et al. (2021).
}
