% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loesstwodRDD.R
\name{loess_calc_point}
\alias{loess_calc_point}
\title{Run loess_calc_one_point across data}
\usage{
loess_calc_point(x, y, radius, data)
}
\arguments{
\item{x}{One side of sentinels.}

\item{y}{Other side of sentinels.}

\item{radius}{Radius size around each point.}

\item{data}{Data to be analyzed.}
}
\description{
Run loess_calc_one_point across data
}
